requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
pydantic
python-dotenv
requests
sentence-transformers
faiss-cpu